Move Logging To Utils
