+++ ModelConverter/render/sgdMdl.h
@@ -2,7 +2,7 @@
 
 #include "render.h"
 #include "game/sgd.h"
-#include "logging.h"
+#include "utils/logging.h"
 #include "math/linalg.h"
 
 void DisplayFF2Model(const char* filename);
@@ -21,5 +21,3 @@ void HandleWeightedMesh(int meshIndex, Vector3& vertex, Vector3& normal);
 void HandleUniqueMesh(int meshIndex, Vector3& vertex, Vector3& normal);
 void HandleTextureMesh(int meshIndex, Vector3& vertex, Vector3& normal);
 SGDCOORDINATE* GetCurrentCoordinate();
-
-
