+++ ModelConverter/utils/logging.cpp
